fix consumer shutdown delay with multiple fetchers per partition
The max time it would take to shut down the consumer was MaxWaitTime
times the number of fetchers ("workers") per partition. Now it is just
MaxWaitTime regardless of the number of partitions or fetchers. This
came up in integration tests of a downstream high-level client, when I
cranked up the number of fetchers on a single partition to 100 ;-)