change producer so it requires a list of partitions to produce to
Previous behavior was that the producer made a Metadata call to get a
list of all partition leaders for the topic, got the partitions, and
wrote to them all. There was a bug in that, because if for whatever
reason there was no leader at the time the call was made, that partition
would never get produced to. So initially I replaced it with just
extracting all partitions, not leaders, from the metadata response.

But then I realized it actually is better for partitions to be specified
by the higher lever client / the user. This is how the consumer does
that. The impact will be moving the partition "discovery" logic to
whatever is using the producer.