bug fix in producer initialization
List of partitions was not ranged through correctly. Didn't notice it
before because the list of partitions for the producer almost always
starts with zero and is monotonic. Came up in tests for partitioned
production (that code coming up shortly, but breaking this out into
separate commit because this is a stand alone bug).