Reset connection if commit returns NOT_COORDINATOR error code
This commit updates kafkaclient in order to correctly react when
the group's coordinator get re-assigned.

It happens, that for some reasons, Group's coordinator can be
re-assigned without notifying the client. In that case, every
client should discover new coordinator and continue communication
with the new one. The change proposed in the commit detects that
coordinator was changed and drops the connection. However, error
NOT_COORDINATOR will be returned but next attempt to commit
offsets will start with discovering the new Coordinator.