have DumbOffsetsManager disconnect from coordinator on any error
DumbOffsetsManager Fetch and Commit methods will now close the
connection to the group coordinator on any error. This is mostly to deal
with situations where the group coordinator has changed, resulting in
error code 15. But, like with things based on the PartitionClient
(producer, consumer), it seems easiest for things based in the
GroupClient to just close (and re-open on next call) without getting
into details of what exactly the error is.

Prerequisite for this was adding Close method to GroupClient in
libkafka.